+++ tiled/client/context.py
@@ -226,10 +226,12 @@ class Context:
             # starlette is available.
             from starlette.testclient import TestClient
 
+            base_uri = f"{uri.scheme}://{uri.netloc}"
             # verify parameter is dropped, as there is no SSL in ASGI mode
             client = TestClient(
                 app=app,
                 raise_server_exceptions=raise_server_exceptions,
+                base_url=base_uri,
             )
             client.timeout = timeout
             client.headers = headers
@@ -444,12 +446,13 @@ class Context:
         timeout=None,
         api_key=UNSET,
         raise_server_exceptions=True,
+        uri=None,
     ):
         """
         Construct a Context around a FastAPI app. Primarily for testing.
         """
         context = cls(
-            uri="http://local-tiled-app/api/v1",
+            uri="http://local-tiled-app/api/v1" if not uri else uri,
             headers=headers,
             api_key=None,
             cache=cache,
@@ -517,7 +520,7 @@ class Context:
                     )
                 ).json()
 
-    def create_api_key(self, scopes=None, expires_in=None, note=None):
+    def create_api_key(self, scopes=None, access_tags=None, expires_in=None, note=None):
         """
         Generate a new API key.
 
@@ -529,6 +532,9 @@ class Context:
         scopes : Optional[List[str]]
             Restrict the access available to the API key by listing specific scopes.
             By default, this will have the same access as the user.
+        access_tags : Optional[List[str]]
+            Restrict the access available to the API key by listing specific tags.
+            By default, this will have no limits on access tags.
         expires_in : Optional[Union[int, str]]
             Number of seconds until API key expires, given as integer seconds
             or a string like: '3y' (years), '3d' (days), '5m' (minutes), '1h'
@@ -545,7 +551,12 @@ class Context:
                     self.http_client.post(
                         self.server_info.authentication.links.apikey,
                         headers={"Accept": MSGPACK_MIME_TYPE},
-                        json={"scopes": scopes, "expires_in": expires_in, "note": note},
+                        json={
+                            "scopes": scopes,
+                            "access_tags": access_tags,
+                            "expires_in": expires_in,
+                            "note": note,
+                        },
                     )
                 ).json()
 
@@ -875,7 +886,9 @@ class Admin:
                     )
                 ).json()
 
-    def create_api_key(self, uuid, scopes=None, expires_in=None, note=None):
+    def create_api_key(
+        self, uuid, scopes=None, access_tags=None, expires_in=None, note=None
+    ):
         """
         Generate a new API key for another user or service.
 
@@ -886,6 +899,9 @@ class Admin:
         scopes : Optional[List[str]]
             Restrict the access available to the API key by listing specific scopes.
             By default, this will have the same access as the principal.
+        access_tags : Optional[List[str]]
+            Restrict the access available to the API key by listing specific tags.
+            By default, this will have no limits on access tags.
         expires_in : Optional[int]
             Number of seconds until API key expires. If None,
             it will never expire or it will have the maximum lifetime
@@ -899,7 +915,12 @@ class Admin:
                     self.context.http_client.post(
                         f"{self.base_url}/auth/principal/{uuid}/apikey",
                         headers={"Accept": MSGPACK_MIME_TYPE},
-                        json={"scopes": scopes, "expires_in": expires_in, "note": note},
+                        json={
+                            "scopes": scopes,
+                            "access_tags": access_tags,
+                            "expires_in": expires_in,
+                            "note": note,
+                        },
                     )
                 ).json()